split associative array implementation from constructor
